+++ src/subcommand/init/shell/bash.rs
@@ -1,0 +1,45 @@
+use super::{posix, HookConfig, ShellConfig};
+
+use anyhow::Result;
+
+use std::borrow::Cow;
+
+pub const CONFIG: ShellConfig = ShellConfig {
+    z: posix::CONFIG.z,
+    alias: posix::CONFIG.alias,
+    hook: HookConfig {
+        prompt: HOOK_PROMPT,
+        pwd: hook_pwd,
+    },
+};
+
+const HOOK_PROMPT: &str = r#"
+_zoxide_hook() {
+    zoxide add
+}
+
+case "$PROMPT_COMMAND" in
+    *_zoxide_hook*) ;;
+    *) PROMPT_COMMAND="_zoxide_hook${PROMPT_COMMAND:+;${PROMPT_COMMAND}}" ;;
+esac
+"#;
+
+const fn hook_pwd() -> Result<Cow<'static, str>> {
+    const HOOK_PWD: &str = r#"
+_zoxide_hook() {
+    if [ -z "${_ZO_PWD}" ]; then
+        _ZO_PWD="${PWD}"
+    elif [ "${_ZO_PWD}" != "${PWD}" ]; then
+        _ZO_PWD="${PWD}"
+        zoxide add
+    fi
+}
+
+case "$PROMPT_COMMAND" in
+    *_zoxide_hook*) ;;
+    *) PROMPT_COMMAND="_zoxide_hook${PROMPT_COMMAND:+;${PROMPT_COMMAND}}" ;;
+esac
+"#;
+
+    Ok(Cow::Borrowed(HOOK_PWD))
+}